+++ packages/utils/src/index.ts
@@ -10,7 +10,7 @@ export { default as checkAppRole } from './checkAppRole';
 export { default as getAppBlocks } from './getAppBlocks';
 export type { BlockMap } from './getAppBlocks';
 export { default as mapValues } from './mapValues';
-export { default as remap, Remapper, Remappers } from './remap';
+export { default as remap } from './remap';
 export {
   default as validateAppDefinition,
   AppsembleValidationError,