+++ CHANGELOG.md
@@ -15,6 +15,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Changed
 
+- App: The `remap` property of actions now uses the new remapper functionality.
 - Server: Converted user IDs to use UUID instead of auto-incremented integers.
 
 ## [0.12.8] - 2020-04-29